requirements.txt
google-cloud-firestore==2.14.0
requests==2.31.0
tmdbsimple==2.9.1
rebrick==0.4.0
algoliasearch==4.36.1
streamlit
google.cloud
firestore
internetarchive
tmdbsimple
fastapi
requests_futures
google-cloud-secret-manager